pyproject.toml
[tool.poetry]
name = "delivery-django-cllient"
version = "0.1.0"
description = "Delivery client backend"
authors = ["Sarawan backend team"]
readme = "README.md"
packages = [{include = "delivery_django_cllient"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
djangorestframework = "3.14.0"
psycopg2 = "^2.9.9"
dynaconf = "^3.2.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
